# swagger.yaml
openapi: 3.0.0
info:
  title: IT Project  Online Cookbook API
  contact: {}
  version: '1.0'
servers:
- url: https://itproject-online-cookbook.herokuapp.com/api/v1
  variables: {}
paths:
...
  /recipe/image:
    post:
      tags:
      - Recipe
      summary: Upload image
      operationId: Uploadimage
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /recipe/image/{id}:
    delete:
      tags:
      - Recipe
      summary: Delete image
      operationId: Deleteimage
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The imageId
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...